backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.0
openai==1.3.0
pyttsx3==2.90
SpeechRecognition==3.10.0
python-dotenv==1.0.0
pydantic==1.10.11
